==> 02_Datenabfrage_Scopus_ALD.ipynb <==
# Datenabfrage von Scopus


# Table of contents

1. [Abfrage Daten Scopus für Zeitscheibe 1: 2010 - 2013](#import_scopus_2010)
2. [Abfrage Daten Scopus für Zeitscheibe 2: 2020 - 2023](#import_scopus_2020)

import pybliometrics
import pandas as pd
from pybliometrics.scopus import ScopusSearch

pybliometrics.scopus.init()  # read API key

import jupyter_black

jupyter_black.load()

%load_ext watermark
%watermark -v --iversions

## Abfrage Daten Scopus für Zeitscheibe 1: 2010 - 2013<a class="anchor" id="import_scopus_2010"></a>


# Datenabfrage fuer ALD Zeitscheibe 1: 2010 - 2013
query2010 = 'TITLE-ABS-Key("atomic layer deposition" OR ("ALD" AND "thin film*" AND NOT ("alcohol-related liver disease"))) AND PUBYEAR > 2009 AND PUBYEAR < 2014'
search2010 = ScopusSearch(query2010, verbose=True, download=True)

print(search2010.get_results_size())

# Daten in DataFrame umwandeln
search2010_df = pd.DataFrame([doc._asdict() for doc in search2010.results])

# Alle Daten abspeichern
search2010_df.to_csv(f"../data/raw/data_scopus_2010.csv", encoding="utf-8", index=False)

### Analyse Daten Scopus für Zeitscheibe 1

# Publikationsarten
type_counts2010 = search2010_df["subtypeDescription"].value_counts()
print(type_counts2010)

conf_review2010 = search2010_df[
    search2010_df["subtypeDescription"] == "Conference Review"
]
pd.options.display.max_colwidth = 200

print(conf_review2010.iloc[54])
conf_review2010["publicationName"].value_counts()

search2010_df[search2010_df["subtypeDescription"] == "Editorial"]

search2010_df[search2010_df["subtypeDescription"] == "Short Survey"]

# keine DOI oder kein Abstract über alle Daten
total_count = search2010_df.shape[0]
print(total_count)
no_doi = search2010_df[search2010_df["doi"].isnull()].shape[0]
no_abstract = search2010_df[search2010_df["description"].isnull()].shape[0]
print(
    f"Keine DOI: {no_doi} ({no_doi/total_count*100:.2f} %) \nkein Abstract: {no_abstract} ({no_abstract/total_count*100:.2f} %)"
)

# Wieviele Einträge davon haben keine DOI bzw. keinen Abstract?

for type in [
    "Article",
    "Conference Paper",
    "Review",
    "Conference Review",
    "Book Chapter",
    "Erratum",
    "Letter",
    "Short Survey",
    "Editorial",
    "Book",
]:
    total_count = type_counts2010[type]
    search2010_type = search2010_df[search2010_df["subtypeDescription"] == type]
    no_doi = search2010_type[search2010_type["doi"].isnull()].shape[0]
    no_abstract = search2010_type[search2010_type["description"].isnull()].shape[0]
    print(
        f"Von {total_count} Einträgen vom Typ {type} haben {no_doi} ({no_doi/total_count*100:.2f} %) Einträge keine DOI und {no_abstract} ({no_abstract/total_count*100:.2f} %) keinen Abstract."
    )

## Abfrage Daten Scopus für Zeitscheibe 2: 2020 - 2023<a class="anchor" id="import_scopus_2020"></a>


# Datenabfrage für ALD
query2020 = 'TITLE-ABS-Key("atomic layer deposition" OR ("ALD" AND "thin film*" AND NOT ("alcohol-related liver disease"))) AND PUBYEAR > 2019 AND PUBYEAR < 2024'
search2020 = ScopusSearch(query2020, verbose=True, download=True)

print(search2020.get_results_size())

# Daten in DataFrame umwandeln
search2020_df = pd.DataFrame([doc._asdict() for doc in search2020.results])
search2020_df.shape

# Alle Daten abspeichern
search2020_df.to_csv(f"../data/raw/data_scopus_2020.csv", encoding="utf-8", index=False)

### Analyse Zeitscheibe 2

# Publikationsarten
type_counts2020 = search2020_df["subtypeDescription"].value_counts()
print(type_counts2020)

search2020_df[search2020_df["subtypeDescription"] == "Note"]

# keine DOI oder kein Abstract über alle Daten
total_count = search2020_df.shape[0]
no_doi = search2020_df[search2020_df["doi"].isnull()].shape[0]
no_abstract = search2020_df[search2020_df["description"].isnull()].shape[0]
print(
    f"Keine DOI: {no_doi} ({no_doi/total_count*100:.2f} %) \nkein Abstract: {no_abstract} ({no_abstract/total_count*100:.2f} %)"
)

# Wieviele Einträge davon haben keine DOI bzw. keinen Abstract?

for type in [
    "Article",
    "Conference Paper",
    "Review",
    "Book Chapter",
    "Erratum",
    "Conference Review",
    "Book",
    "Note",
    "Editorial",
    "Letter",
    "Short Survey",
    "Retracted",
    "Data Paper",
]:
    total_count = type_counts2020[type]
    search2020_type = search2020_df[search2020_df["subtypeDescription"] == type]
    no_doi = search2020_type[search2020_type["doi"].isnull()].shape[0]
    no_abstract = search2020_type[search2020_type["description"].isnull()].shape[0]
    print(
        f"Von {total_count} Einträgen vom Typ {type} haben {no_doi} ({no_doi/total_count*100:.2f} %) Einträge keine DOI und {no_abstract} ({no_abstract/total_count*100:.2f} %) keinen Abstract."
    )
