==> 05_Matching_Scopus_OpenAlex.ipynb <==
# Matching der Ergebnisse Scopus und OpenAlex


## Inhalt

* [Zeitscheibe 1: 2010 - 2013](#2010)
  * [Matching](#matching_2010)
* [Zeitscheibe 2: 2020 - 2023](#2020)
  * [Matching](#matching_2020)

import pandas as pd
import numpy as np
from rapidfuzz.distance import Levenshtein
import uuid
import matplotlib.pyplot as plt

pd.options.mode.copy_on_write = True

import jupyter_black

jupyter_black.load()

%load_ext watermark
%watermark -v --iversions

# implementation based on https://github.com/gavieira/biblioverlap/blob/main/R/04-score_matching.R


def calc_distance_score_matrix(db1_col, db2_col, max_score, penalty):
    """  Compute normalized Levenshtein distance matrix. """ 
    n, m = len(db1_col), len(db2_col)
    matrix = np.zeros((n, m), dtype=np.float32)
    for i, val1 in enumerate(db1_col):
        for j, val2 in enumerate(db2_col):
            dist = Levenshtein.normalized_distance(str(val1), str(val2))
            matrix[i, j] = max_score - dist * penalty
    return matrix


def calc_year_score_matrix(db1_col, db2_col):
    """  Compute score for year  difference. """
    n, m = len(db1_col), len(db2_col)
    matrix = np.zeros((n, m), dtype=np.float32)
    for i, val1 in enumerate(db1_col):
        for j, val2 in enumerate(db2_col):
            if val1 == val2:
                matrix[i, j] = 0
            elif abs(val1 - val2) == 1:
                matrix[i, j] = 0.5
            else:
                matrix[i, j] = 1.0

    return matrix


def extract_score_matches(final_score_matrix, db1_index, db2_index):
    """ Get matches based on score. """
    max_scores = final_score_matrix.max(
        axis=1
    )  # get max value for all rows (db1 values)
    db2_pos = final_score_matrix.argmax(axis=1)
    matches = []
    for db1_id, score, db2_idx in zip(db1_index, max_scores, db2_pos):
        if score != 0:
            matches.append(
                {"scopus_id": db1_id, "score": score, "openalex_id": db2_index[db2_idx]}
            )
    return matches


def score_matching(
    db1,
    db2,
    title_penalty,
    title_max,
    source_penalty,
    source_max,
    author_penalty,
    author_max,
    score_cutoff,
):
    """
    Conduct score based matching.

    Parameters
    ----------
    db1 : pandas.DataFrame
        First collection of works for comparison, from Scopus.
    db2 : pandas.DataFrame
        Second collection of works for comparison, from OpenAlex.
    title_penalty: float
        Penalty for title score calculation
    title_max: float
        Maximum value for title score calculation
    source_penalty: float
        Penalty for source title score calculation
    source_max: float
        Maximum value for source title score calculation
    author_penalty : float
        Penalty for author name score calculation
    author_max : float
        Maximum value for author name score calculation
    score_cutoff : float
        Cutoff value to determine matching
     
    Returns
    --------
    matches : array of dict
        Found matches, contains scopus_id from db1, score and openalex_id from db2
    final_score_matrix : np.array
        Matrix containing final score values.
    """ 
    title_matrix = calc_distance_score_matrix(
        db1["scopus_title"], db2["openalex_title"], title_max, title_penalty
    )
    author_matrix = calc_distance_score_matrix(
        db1["scopus_name_1st_author"],
        db2["openalex_name_1st_author"],
        author_max,
        author_penalty,
    )
    source_matrix = calc_distance_score_matrix(
        db1["scopus_source_title"],
        db2["openalex_source_title"],
        source_max,
        source_penalty,
    )
    year_matrix = calc_year_score_matrix(
        db1["scopus_publication_year"], db2["openalex_publication_year"]
    )

    final_score_matrix = title_matrix + author_matrix + source_matrix - year_matrix
    final_score_matrix = final_score_matrix * (final_score_matrix >= score_cutoff)
    matches = extract_score_matches(
        final_score_matrix, db1["scopus_id"], db2["openalex_id"]
    )
    return matches, final_score_matrix

# Check: Wurden Anhand des Score Publikationen gematcht, welche unterschiedliche DOIS besitzen? 
# --> Test für mögliche falsch positive Matches


def get_scopus_id_for_false_positive_matches(data_matches, data_scopus, data_openalex):
    """
    Check found matches and identify falsely matched works.

    Parameters
    ----------
    data_matches : array of dict
        Found matches, contains scopus_id from db1, score and openalex_id from db2
    data_scopus : pandas.DataFrame
        Collections of works from Scopus.
    data_openalex : pandas.DataFrame
        Collections of works from OpenAlex.
     
    Returns
    --------
    index_list_to_be_removed : list
        Indexes of data_matches which contain falsely matched works.
    """
    index_list_to_be_removed = []
    data_matches_df = pd.DataFrame.from_dict(data_matches)

    for index, row in data_matches_df.iterrows():
        scopus_id = row["scopus_id"]
        openalex_id = row["openalex_id"]
        scopus_data = data_scopus[data_scopus["scopus_id"] == scopus_id]
        openalex_data = data_openalex[data_openalex["openalex_id"] == openalex_id]

        if not scopus_data.empty:  # Scopus Eintrag vorhanden?
            if not openalex_data.empty:  # OpenAlex Eintrag vorhanden?
                scopus_doi = scopus_data.iloc[0, scopus_data.columns.get_loc("doi")]
                openalex_doi = openalex_data.iloc[
                    0, openalex_data.columns.get_loc("doi")
                ]
                if scopus_doi != openalex_doi:  #  DOI unterschiedlich?
                    index_list_to_be_removed.append(index)
    return index_list_to_be_removed


def remove_matches_with_different_doi(data_matches, data_scopus, data_openalex):
    """
    Check found matches and remove falsely matched works.

    Parameters
    ----------
    data_matches : array of dict
        Found matches, contains scopus_id from db1, score and openalex_id from db2
    data_scopus : pandas.DataFrame
        Collections of works from Scopus.
    data_openalex : pandas.DataFrame
        Collections of works from OpenAlex.
     
    Returns
    --------
    data_matches_update : array of dict
        Updated array of matches with falsely matched works removed.
    """
    index_to_be_removed = get_scopus_id_for_false_positive_matches(
        data_matches, data_scopus, data_openalex
    )
    data_matches_update = [
        i for j, i in enumerate(data_matches) if j not in index_to_be_removed
    ]
    return data_matches_update

## Zeitscheibe 1: 2010 - 2013<a class="anchor" id="2010"></a>


# Daten einlesen
data_scopus_2010 = pd.read_csv("../data/processed/data_scopus_2010_cleaned.csv")
data_openalex_2010 = pd.read_csv("../data/processed/data_openalex_2010_cleaned.csv")

data_scopus_2010.columns

# Spaltennamen umbenennen für Abgleich
data_scopus_2010.rename(
    columns={
        "eid": "scopus_id",
        "title": "scopus_title",
        "name_1st_author": "scopus_name_1st_author",
        "publicationName": "scopus_source_title",
        "volume": "scopus_volume",
        "publication_year": "scopus_publication_year",
    },
    inplace=True,
)
data_openalex_2010.rename(
    columns={
        "id": "openalex_id",
        "title": "openalex_title",
        "name_1st_author": "openalex_name_1st_author",
        "source_title": "openalex_source_title",
        "volume": "openalex_volume",
        "publication_year": "openalex_publication_year",
    },
    inplace=True,
)

data_scopus_2010.head()

data_openalex_2010.columns

data_openalex_2010.index[-1]

data_openalex_2010.head()

### Matching<a class="anchor" id="matching_2010"></a>

Das Vorgehen zum Abgleich der Datenbanken bzw. Matching der Datenbankeinträge orientiert sich an [Vieira & Leta, 2024](https://doi.org/10.1007/s11192-024-05065-5), welche die documentbasierte Matching-Stratgie der R-Bibliothek *biblioverlap* beschreibt. 

# Erzeuge UUID zur eindeutigen Identifikation der Daten
# Scopus dient als UUID Spender, bei erfolgten Match wird der OpenAlex Datensatz mit dieser UUID verknüpft

data_scopus_2010["uuid"] = [uuid.uuid4() for _ in range(len(data_scopus_2010.index))]
data_openalex_2010["uuid"] = pd.Series(dtype="object")

data_scopus_2010["uuid"]

data_openalex_2010["uuid"]

# Desweiteren wird ein Label genutzt um schon gematchte Datensätze besser identifizieren zu können
data_scopus_2010["matched"] = False
data_openalex_2010["matched"] = False

# Schritt 1: Matching anhand der DOI
# Zuerst werden die Einträge mit DOI aus Scopus mit den Einträgen mit DOI aus OpenAlex verglichen.

# Nur Datensätze mit DOI berücksichtigen
data_scopus_2010_mitdoi = data_scopus_2010[data_scopus_2010["doi"].notnull()]
data_openalex_2010_mitdoi = data_openalex_2010[data_openalex_2010["doi"].notnull()]

# Index aktualisieren
data_scopus_2010_mitdoi.reset_index(drop=True, inplace=True)
data_openalex_2010_mitdoi.reset_index(drop=True, inplace=True)

print(data_scopus_2010_mitdoi.shape[0])
print(data_openalex_2010_mitdoi.shape[0])

for index_scopus, row_scopus in data_scopus_2010_mitdoi.iterrows():
    for index_openalex, row_openalex in data_openalex_2010_mitdoi.iterrows():
        if row_scopus["doi"] == row_openalex["doi"]:
            # mark rows as matched
            data_scopus_2010_mitdoi.iloc[
                index_scopus, data_scopus_2010_mitdoi.columns.get_loc("matched")
            ] = True
            data_openalex_2010_mitdoi.iloc[
                index_openalex, data_openalex_2010_mitdoi.columns.get_loc("matched")
            ] = True
            # write UUID to OpenAlex
            data_openalex_2010_mitdoi.iloc[
                index_openalex, data_openalex_2010_mitdoi.columns.get_loc("uuid")
            ] = row_scopus["uuid"]

# No. of matched items
data_scopus_2010_doi_matched = data_scopus_2010_mitdoi[
    data_scopus_2010_mitdoi["matched"] == True
]
data_openalex_2010_doi_matched = data_openalex_2010_mitdoi[
    data_openalex_2010_mitdoi["matched"] == True
]
num_scopus = data_scopus_2010_doi_matched.shape[0]
num_openalex = data_openalex_2010_doi_matched.shape[0]

if num_scopus == num_openalex:
    print(f"Es wurden {num_scopus} übereinstimmende DOIs gefunden.")

Für alle weiteren Datensätze wird in einem zweiten Schritt ein Matching anhand der Metadaten erzielt.   
Die Entwicklung des genutzen Scores erfolgt in einem separaten Notebook ('05a_Matching_score.ipynb'). Hierfür wurden die per DOI gematchten Daten für die Zeitscheibe 2010 - 2013 genutzt.

# scopus data
data_doi_matched_scopus = data_scopus_2010_doi_matched[
    [
        "eid",
        "doi",
        "title",
        "name_1st_author",
        "publicationName",
        "volume",
        "publication_year",
    ]
].copy()
data_doi_matched_scopus.rename(
    columns={
        "eid": "scopus_id",
        "title": "scopus_title",
        "name_1st_author": "scopus_name_1st_author",
        "publicationName": "scopus_source_title",
        "volume": "scopus_volume",
        "publication_year": "scopus_publication_year",
    },
    inplace=True,
)

# openalex data
data_doi_matched_openalex = data_openalex_2010_doi_matched[
    [
        "id",
        "doi",
        "title",
        "name_1st_author",
        "source_title",
        "volume",
        "publication_year",
    ]
].copy()
data_doi_matched_openalex.rename(
    columns={
        "id": "openalex_id",
        "title": "openalex_title",
        "name_1st_author": "openalex_name_1st_author",
        "source_title": "openalex_source_title",
        "volume": "openalex_volume",
        "publication_year": "openalex_publication_year",
    },
    inplace=True,
)

# join
data_doi_matched = pd.merge(
    data_doi_matched_scopus, data_doi_matched_openalex, on="doi"
)

# save data
data_doi_matched.to_csv("../data/processed/data_doi_matched.csv", index=False)

# Erstellung der noch zu matchenden Datensätze: ohne DOI und die mit DOI aber unmatched

data_scopus_2010_doi_unmatched = data_scopus_2010_mitdoi[
    data_scopus_2010_mitdoi["matched"] == False
]
data_openalex_2010_doi_unmatched = data_openalex_2010_mitdoi[
    data_openalex_2010_mitdoi["matched"] == False
]

data_scopus_2010_nodoi = data_scopus_2010[data_scopus_2010["doi"].isnull()]
data_openalex_2010_nodoi = data_openalex_2010[data_openalex_2010["doi"].isnull()]

data_scopus_2010_unmatched = pd.concat(
    [data_scopus_2010_doi_unmatched, data_scopus_2010_nodoi], ignore_index=True
)
data_openalex_2010_unmatched = pd.concat(
    [data_openalex_2010_doi_unmatched, data_openalex_2010_nodoi], ignore_index=True
)

print(
    f"Für Scopus sind {data_scopus_2010.shape[0]} Einträge vorhanden. Hiervon haben {data_scopus_2010_mitdoi.shape[0]} eine DOI und {data_scopus_2010_nodoi.shape[0]} keine DOI."
)
print(
    f"Für OpenAlex sind {data_openalex_2010.shape[0]} Einträge vorhanden. Hiervon haben {data_openalex_2010_mitdoi.shape[0]} eine DOI und {data_openalex_2010_nodoi.shape[0]} keine DOI."
)

# ohne Titel
print(
    f"Scopus ohne Titel: {data_scopus_2010[data_scopus_2010['scopus_title'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Titel: {data_openalex_2010[data_openalex_2010['openalex_title'].isnull()].shape[0]} \n"
)

# ohne Publikationsjahr

print(
    f"Scopus ohne Publikationsjahr: {data_scopus_2010[data_scopus_2010['scopus_publication_year'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Publikationsjahr: {data_openalex_2010[data_openalex_2010['openalex_publication_year'].isnull()].shape[0]} \n"
)

# ohne Autornamen

print(
    f"Scopus ohne Autorennamen: {data_scopus_2010[data_scopus_2010['scopus_name_1st_author'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Autorennamen: {data_openalex_2010[data_openalex_2010['openalex_name_1st_author'].isnull()].shape[0]} \n"
)

# ohne Titel Quelle (e.g. Journal)

print(
    f"Scopus ohne Quellentitel: {data_scopus_2010[data_scopus_2010['scopus_source_title'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Quellentitel: {data_openalex_2010[data_openalex_2010['openalex_source_title'].isnull()].shape[0]} \n"
)

# ohne Volume

print(
    f"Scopus ohne Ausgabe (Volume): {data_scopus_2010[data_scopus_2010['scopus_volume'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Ausgabe (Volume): {data_openalex_2010[data_openalex_2010['openalex_volume'].isnull()].shape[0]}\n"
)


# ohne ISSN

print(
    f"Scopus ohne ISSN: {data_scopus_2010[data_scopus_2010['issn'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne ISSN: {data_openalex_2010[data_openalex_2010['source_issn'].isnull()].shape[0]}\n"
)


# ohne Seitenangaben

print(
    f"Scopus ohne Seitenangaben: {data_scopus_2010[data_scopus_2010['pageRange'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Seitenangaben: {data_openalex_2010[data_openalex_2010['first_page'].isnull()].shape[0]}\n"
)

# ohne Abstract

print(
    f"Scopus ohne Abstract: {data_scopus_2010[data_scopus_2010['description'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Abstract: {data_openalex_2010[data_openalex_2010['abstract'].isnull()].shape[0]}"
)

print(data_scopus_2010_unmatched.shape[0])
print(data_openalex_2010_unmatched.shape[0])

print(
    data_scopus_2010_unmatched[data_scopus_2010_unmatched["matched"] == True].shape[0]
)
print(
    data_openalex_2010_unmatched[data_openalex_2010_unmatched["matched"] == True].shape[
        0
    ]
)

data_scopus_2010_unmatched

# match rest of data
matches_2010, score_matrix_2010 = score_matching(
    data_scopus_2010_unmatched,
    data_openalex_2010_unmatched,
    title_penalty=0.1,
    title_max=0.5,
    source_penalty=0.1,
    source_max=0.3,
    author_penalty=0.1,
    author_max=0.3,
    score_cutoff=1.0,
)

# Entferne Matches mit unterschiedlicher DOI (vermutlich falsch positiv)
matches_2010 = remove_matches_with_different_doi(
    matches_2010, data_scopus_2010_mitdoi, data_openalex_2010_mitdoi
)

matches_2010_df = pd.DataFrame.from_dict(matches_2010)
matches_2010_df

matches_2010_df[matches_2010_df.duplicated("scopus_id", keep=False)]

matches_2010_df[matches_2010_df.duplicated("openalex_id", keep=False)]

# copy results back to dataframes
data_openalex_2010_unmatched["score"] = None
for match in matches_2010:
    # set entries as matched
    index_scopus = data_scopus_2010_unmatched[
        data_scopus_2010_unmatched["scopus_id"] == match["scopus_id"]
    ].index.to_list()[0]
    index_openalex = data_openalex_2010_unmatched[
        data_openalex_2010_unmatched["openalex_id"] == match["openalex_id"]
    ].index.to_list()[0]

    data_scopus_2010_unmatched.iloc[
        index_scopus, data_scopus_2010_unmatched.columns.get_loc("matched")
    ] = True
    data_openalex_2010_unmatched.iloc[
        index_openalex, data_openalex_2010_unmatched.columns.get_loc("matched")
    ] = True

    # write UUID to OpenAlex
    uuid_scopus = data_scopus_2010_unmatched.iloc[
        index_scopus, data_scopus_2010_unmatched.columns.get_loc("uuid")
    ]
    data_openalex_2010_unmatched.iloc[
        index_openalex, data_openalex_2010_unmatched.columns.get_loc("uuid")
    ] = uuid_scopus

    # write score to openalex data
    data_openalex_2010_unmatched.iloc[
        index_openalex, data_openalex_2010_unmatched.columns.get_loc("score")
    ] = match["score"]

data_scopus_2010_unmatched[data_scopus_2010_unmatched["matched"] == True].shape[0]

data_scopus_2010_unmatched[data_scopus_2010_unmatched["matched"] == False].shape[0]

data_openalex_2010_unmatched[data_openalex_2010_unmatched["matched"] == False].shape[0]

# No. of matched items
data_scopus_2010_score_matched = data_scopus_2010_unmatched[
    data_scopus_2010_unmatched["matched"] == True
]
data_openalex_2010_score_matched = data_openalex_2010_unmatched[
    data_openalex_2010_unmatched["matched"] == True
]
num_scopus = data_scopus_2010_score_matched.shape[0]
num_openalex = data_openalex_2010_score_matched.shape[0]

if num_scopus == num_openalex:
    print(f"Es wurden {num_scopus} übereinstimmende DOIs gefunden.")

# entries in openalex with same uuid??
data_openalex_matched = data_openalex_2010_unmatched[
    data_openalex_2010_unmatched["matched"] == True
]
data_openalex_matched[data_openalex_matched.duplicated("uuid", keep=False)]

# Ergebnisse aus beiden Matching Schritten zusammen führen

data_scopus_2010_matched = pd.concat(
    [data_scopus_2010_doi_matched, data_scopus_2010_score_matched]
)
data_openalex_2010_matched = pd.concat(
    [data_openalex_2010_doi_matched, data_openalex_2010_score_matched]
)

data_2010_matched = pd.merge(
    data_scopus_2010_matched, data_openalex_2010_matched, on="uuid"
)

# Spalten umbenennen
data_2010_matched.rename(
    columns={
        "matched_y": "matched",
        "doi_x": "doi",
    },
    inplace=True,
)

# Spalte uuid nicht mit abspeichern
data_2010_matched.drop(columns=["uuid", "matched_x", "doi_y"], inplace=True)

data_2010_matched.columns

# Daten speichern
data_2010_matched.to_csv("../data/diff/data_2010_matched.csv")

data_scopus_2010_unmatched[data_scopus_2010_unmatched["matched"] == False].to_csv(
    "../data/diff/data_2010_only_scopus.csv"
)
data_openalex_2010_unmatched[data_openalex_2010_unmatched["matched"] == False].to_csv(
    "../data/diff/data_2010_only_openalex.csv"
)

# Daten für Veröffentlichung (nur DOI, EID und ID)
df_2010_matched = data_2010_matched[
    [
        "doi",
        "scopus_id",
        "openalex_id",
    ]
]
df_2010_only_scopus = data_scopus_2010_unmatched[
    [
        "doi",
        "scopus_id",
    ]
]
df_2010_only_openalex = data_openalex_2010_unmatched[
    [
        "doi",
        "openalex_id",
    ]
]

df_2010_matched_all = pd.concat(
    [df_2010_matched, df_2010_only_scopus, df_2010_only_openalex]
)

# check: correct df size
df_2010_matched_all.shape[0] == (
    df_2010_matched.shape[0]
    + df_2010_only_scopus.shape[0]
    + df_2010_only_openalex.shape[0]
)

# Daten speichern
df_2010_matched_all.to_csv("../data/diff/data_2010_matching_results.csv")

## Zeitscheibe 2: 2020 - 2023<a class="anchor" id="2020"></a>

# Daten einlesen
data_scopus_2020 = pd.read_csv("../data/processed/data_scopus_2020_cleaned.csv")
data_openalex_2020 = pd.read_csv("../data/processed/data_openalex_2020_cleaned.csv")

# Spaltennamen umbennnen für Abgleich
data_scopus_2020.rename(
    columns={
        "eid": "scopus_id",
        "title": "scopus_title",
        "name_1st_author": "scopus_name_1st_author",
        "publicationName": "scopus_source_title",
        "volume": "scopus_volume",
        "publication_year": "scopus_publication_year",
    },
    inplace=True,
)
data_openalex_2020.rename(
    columns={
        "id": "openalex_id",
        "title": "openalex_title",
        "name_1st_author": "openalex_name_1st_author",
        "source_title": "openalex_source_title",
        "volume": "openalex_volume",
        "publication_year": "openalex_publication_year",
    },
    inplace=True,
)

data_scopus_2020.head()

data_openalex_2020.head()

### Matching<a class="anchor" id="matching_2020"></a>

# Erzeuge UUID zur eindeutigen Identifikation der Daten
# Scopus dient als UUID Spender, bei erfolgten Match wird der OpenAlex Datensatz mit dieser UUId verknüpft

data_scopus_2020["uuid"] = [uuid.uuid4() for _ in range(len(data_scopus_2020.index))]
data_openalex_2020["uuid"] = pd.Series(dtype="object")

# Desweiteren wird ein Label genutzt um schon gematchte Datensätze besser identifizieren zu können
data_scopus_2020["matched"] = False
data_openalex_2020["matched"] = False

# Aufteilung in Daten mit und ohne DOI
data_scopus_2020_nodoi = data_scopus_2020[data_scopus_2020["doi"].isnull()]
data_scopus_2020_mitdoi = data_scopus_2020[data_scopus_2020["doi"].notnull()]

data_openalex_2020_nodoi = data_openalex_2020[data_openalex_2020["doi"].isnull()]
data_openalex_2020_mitdoi = data_openalex_2020[data_openalex_2020["doi"].notnull()]

print(
    f"Für Scopus wurden {data_scopus_2020.shape[0]} Einträge gefunden. Hiervon haben {data_scopus_2020_mitdoi.shape[0]} eine DOI und {data_scopus_2020_nodoi.shape[0]} keine DOI."
)
print(
    f"Für OpenAlex wurden {data_openalex_2020.shape[0]} Einträge gefunden. Hiervon haben {data_openalex_2020_mitdoi.shape[0]} eine DOI und {data_openalex_2020_nodoi.shape[0]} keine DOI."
)

# ohen Titel
print(
    f"Scopus ohne Titel: {data_scopus_2020[data_scopus_2020['scopus_title'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Titel: {data_openalex_2020[data_openalex_2020['openalex_title'].isnull()].shape[0]} \n"
)

# ohne Publikationsjahr

print(
    f"Scopus ohne Publikationsjahr: {data_scopus_2020[data_scopus_2020['scopus_publication_year'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Publikationsjahr: {data_openalex_2020[data_openalex_2020['openalex_publication_year'].isnull()].shape[0]} \n"
)

# ohne Autornamen

print(
    f"Scopus ohne Autorennamen: {data_scopus_2020[data_scopus_2020['scopus_name_1st_author'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Autorennamen: {data_openalex_2020[data_openalex_2020['openalex_name_1st_author'].isnull()].shape[0]} \n"
)

# ohne Titel Quelle (e.g. Journal)

print(
    f"Scopus ohne Quellentitel: {data_scopus_2020[data_scopus_2020['scopus_source_title'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Quellentitel: {data_openalex_2020[data_openalex_2020['openalex_source_title'].isnull()].shape[0]} \n"
)

# ohne Volume

print(
    f"Scopus ohne Ausgabe (Volume): {data_scopus_2020[data_scopus_2020['scopus_volume'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Ausgabe (Volume): {data_openalex_2020[data_openalex_2020['openalex_volume'].isnull()].shape[0]}\n"
)


# ohne ISSN

print(
    f"Scopus ohne ISSN: {data_scopus_2020[data_scopus_2020['issn'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne ISSN: {data_openalex_2020[data_openalex_2020['source_issn'].isnull()].shape[0]}\n"
)


# ohne Seitenangaben

print(
    f"Scopus ohne Seitenangaben: {data_scopus_2020[data_scopus_2020['pageRange'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Seitenangaben: {data_openalex_2020[data_openalex_2020['first_page'].isnull()].shape[0]}\n"
)

# ohne Abstract

print(
    f"Scopus ohne Abstract: {data_scopus_2020[data_scopus_2020['description'].isnull()].shape[0]}"
)
print(
    f"OpenAlex ohne Abstract: {data_openalex_2020[data_openalex_2020['abstract'].isnull()].shape[0]}"
)

# Schritt 1: Matching anhand der DOI
# Zuerst werden die Einträge mit DOI aus Scopus mit den Einträgen mit DOI aus OpenAlex verglichen.

# Nur Datensätze mit DOI berücksichtigen
# Index aktualisieren
data_scopus_2020_mitdoi.reset_index(drop=True, inplace=True)
data_openalex_2020_mitdoi.reset_index(drop=True, inplace=True)

print(data_scopus_2020_mitdoi.shape[0])
print(data_openalex_2020_mitdoi.shape[0])

for index_scopus, row_scopus in data_scopus_2020_mitdoi.iterrows():
    for index_openalex, row_openalex in data_openalex_2020_mitdoi.iterrows():
        if row_scopus["doi"] == row_openalex["doi"]:
            # mark rows as matched
            data_scopus_2020_mitdoi.iloc[
                index_scopus, data_scopus_2020_mitdoi.columns.get_loc("matched")
            ] = True
            data_openalex_2020_mitdoi.iloc[
                index_openalex, data_openalex_2020_mitdoi.columns.get_loc("matched")
            ] = True
            # write UUID to OpenAlex
            data_openalex_2020_mitdoi.iloc[
                index_openalex, data_openalex_2020_mitdoi.columns.get_loc("uuid")
            ] = row_scopus["uuid"]

# No. of matched items
data_scopus_2020_doi_matched = data_scopus_2020_mitdoi[
    data_scopus_2020_mitdoi["matched"] == True
]
data_openalex_2020_doi_matched = data_openalex_2020_mitdoi[
    data_openalex_2020_mitdoi["matched"] == True
]
num_scopus = data_scopus_2020_doi_matched.shape[0]
num_openalex = data_openalex_2020_doi_matched.shape[0]

if num_scopus == num_openalex:
    print(f"Es wurden {num_scopus} übereinstimmende DOIs gefunden.")

# Schritt 2: Matching anhand des Scores
# Erstellung der noch zu matchenden Datensätze: ohne DOI und die mit DOI aber unmatched

data_scopus_2020_doi_unmatched = data_scopus_2020_mitdoi[
    data_scopus_2020_mitdoi["matched"] == False
]
data_openalex_2020_doi_unmatched = data_openalex_2020_mitdoi[
    data_openalex_2020_mitdoi["matched"] == False
]

data_scopus_2020_nodoi = data_scopus_2020[data_scopus_2020["doi"].isnull()]
data_openalex_2020_nodoi = data_openalex_2020[data_openalex_2020["doi"].isnull()]

data_scopus_2020_unmatched = pd.concat(
    [data_scopus_2020_doi_unmatched, data_scopus_2020_nodoi], ignore_index=True
)
data_openalex_2020_unmatched = pd.concat(
    [data_openalex_2020_doi_unmatched, data_openalex_2020_nodoi], ignore_index=True
)

# match rest of data
matches_2020, score_matrix_2020 = score_matching(
    data_scopus_2020_unmatched,
    data_openalex_2020_unmatched,
    title_penalty=0.1,
    title_max=0.5,
    source_penalty=0.1,
    source_max=0.3,
    author_penalty=0.1,
    author_max=0.3,
    score_cutoff=1.0,
)

# Entferne Matches mit unterschiedlicher DOI (vermutlich falsch positiv)
matches_2020 = remove_matches_with_different_doi(
    matches_2020, data_scopus_2020_mitdoi, data_openalex_2020_mitdoi
)

matches_2020_df = pd.DataFrame.from_dict(matches_2020)
matches_2020_df.head()

matches_2020_df.shape[0]

matches_2020_df[matches_2020_df.duplicated("scopus_id", keep=False)]

matches_2020_df[matches_2020_df.duplicated("openalex_id", keep=False)]

# OpenAlex-Eintrag doppelt gematcht, Eintrag immer geringerem Score entfernen
del matches_2020[24]

# copy results back to dataframes
data_openalex_2020_unmatched["score"] = None
for match in matches_2020:
    # set entries as matched
    index_scopus = data_scopus_2020_unmatched[
        data_scopus_2020_unmatched["scopus_id"] == match["scopus_id"]
    ].index.to_list()[0]
    index_openalex = data_openalex_2020_unmatched[
        data_openalex_2020_unmatched["openalex_id"] == match["openalex_id"]
    ].index.to_list()[0]

    data_scopus_2020_unmatched.iloc[
        index_scopus, data_scopus_2020_unmatched.columns.get_loc("matched")
    ] = True
    data_openalex_2020_unmatched.iloc[
        index_openalex, data_openalex_2020_unmatched.columns.get_loc("matched")
    ] = True

    # write UUID to OpenAlex
    uuid_scopus = data_scopus_2020_unmatched.iloc[
        index_scopus, data_scopus_2020_unmatched.columns.get_loc("uuid")
    ]
    data_openalex_2020_unmatched.iloc[
        index_openalex, data_openalex_2020_unmatched.columns.get_loc("uuid")
    ] = uuid_scopus

    # write score to openalex data
    data_openalex_2020_unmatched.iloc[
        index_openalex, data_openalex_2020_unmatched.columns.get_loc("score")
    ] = match["score"]

data_scopus_2020_unmatched[data_scopus_2020_unmatched["matched"] == False].shape[0]

data_scopus_2020_unmatched[data_scopus_2020_unmatched["matched"] == True].shape[0]

data_openalex_2020_unmatched[data_openalex_2020_unmatched["matched"] == True].shape[0]

# No. of matched items
data_scopus_2020_score_matched = data_scopus_2020_unmatched[
    data_scopus_2020_unmatched["matched"] == True
]
data_openalex_2020_score_matched = data_openalex_2020_unmatched[
    data_openalex_2020_unmatched["matched"] == True
]
num_scopus = data_scopus_2020_score_matched.shape[0]
num_openalex = data_openalex_2020_score_matched.shape[0]

print(num_scopus)
print(num_openalex)

if num_scopus == num_openalex:
    print(f"Es wurden {num_scopus} übereinstimmende DOIs gefunden.")

# Ergebnisse aus beiden Matching-Schritten zusammen führen
data_scopus_2020_matched = pd.concat(
    [data_scopus_2020_doi_matched, data_scopus_2020_score_matched]
)
data_openalex_2020_matched = pd.concat(
    [data_openalex_2020_doi_matched, data_openalex_2020_score_matched]
)

data_2020_matched = pd.merge(
    data_scopus_2020_matched, data_openalex_2020_matched, on="uuid"
)

# Spalten umbenennen
data_2020_matched.rename(
    columns={
        "doi_x": "doi",
        "matched_y": "matched",
    },
    inplace=True,
)

# Spalte UUID nicht mit abspeichern
data_2020_matched.drop(columns=["uuid", "matched_x", "doi_y"], inplace=True)

data_2020_matched.columns

# Daten speichern
data_2020_matched.to_csv("../data/diff/data_2020_matched.csv")

data_scopus_2020_unmatched[data_scopus_2020_unmatched["matched"] == False].to_csv(
    "../data/diff/data_2020_only_scopus.csv"
)
data_openalex_2020_unmatched[data_openalex_2020_unmatched["matched"] == False].to_csv(
    "../data/diff/data_2020_only_openalex.csv"
)

data_scopus_2020_unmatched.columns

# Daten für Veröffentlichung (nur DOI, EID und ID)
df_2020_matched = data_2020_matched[
    [
        "doi",
        "scopus_id",
        "openalex_id",
    ]
]
df_2020_only_scopus = data_scopus_2020_unmatched[
    [
        "doi",
        "scopus_id",
    ]
]
df_2020_only_openalex = data_openalex_2020_unmatched[
    [
        "doi",
        "openalex_id",
    ]
]

df_2020_matched_all = pd.concat(
    [df_2020_matched, df_2020_only_scopus, df_2020_only_openalex]
)

# check: correct df size
df_2020_matched_all.shape[0] == (
    df_2020_matched.shape[0]
    + df_2020_only_scopus.shape[0]
    + df_2020_only_openalex.shape[0]
)

df_2020_matched_all.head(10)

# Daten speichern
df_2020_matched_all.to_csv("../data/diff/data_2020_matching_results.csv")
